✨Added tip_header property to tips table

+++ mephisto/operations/operator.py
@@ -607,34 +607,55 @@ class Operator:
 
         task_name = run_config.task.get("task_name", None)
         tasks = self.db.find_tasks(task_name=task_name)
-        print("task_name:" + tasks[0].task_name)
+        # The first time runnung a task there will be nothing printed as len(tasks) == 0
+        if len(tasks) > 0:
+            print(tasks)
+            print("task_name:" + tasks[0].task_name)
 
-        self.db.new_tip(task_name=task_name, tip_text="this is tip 1")
-        self.db.new_tip(task_name=task_name, tip_text="this is tip 2")
-        self.db.new_tip(task_name=task_name, tip_text="this is tip 3")
-        self.db.new_tip(task_name=task_name, tip_text="this is tip 4")
+            self.db.new_tip(
+                task_name=task_name,
+                tip_header="this is header 1",
+                tip_text="this is tip 1",
+            )
+            self.db.new_tip(
+                task_name=task_name,
+                tip_header="this is header 2",
+                tip_text="this is tip 2",
+            )
+            self.db.new_tip(
+                task_name=task_name,
+                tip_header="this is header 3",
+                tip_text="this is tip 3",
+            )
+            self.db.new_tip(
+                task_name=task_name,
+                tip_header="this is header 4",
+                tip_text="this is tip 4",
+            )
 
-        # Gets tips by task name
-        tips = self.db.get_tip_by_task_name(task_name=task_name)
+            # Gets tips by task name
+            tips = self.db.get_tip_by_task_name(task_name=task_name)
 
-        print("Listing out all tips under " + task_name + ":")
-        for i in range(len(tips)):
-            print("tip_id:" + tips[i].db_id)
-            print("task_name: " + tips[i].task_name)
-            print("tip_text: " + tips[i].tip_text)
+            print("Listing out all tips under " + task_name + ":")
+            for i in range(len(tips)):
+                print("tip_id:" + tips[i].db_id)
+                print("task_name: " + tips[i].task_name)
+                print("tip_header: " + tips[i].tip_header)
+                print("tip_text: " + tips[i].tip_text)
+                print("-----------------")
+
+            second_tip_row = self.db.get_tip(tip_id="2")
+
+            # Creaing a Tip object
+            second_tip = Tip(self.db, str(second_tip_row["tip_id"]), second_tip_row)
+
+            print("getting second tip by id:")
+            print("tip_id:" + second_tip.db_id)
+            print("task_name: " + second_tip.task_name)
+            print("tip_header: " + second_tip.tip_header)
+            print("tip_text: " + second_tip.tip_text)
             print("-----------------")
 
-        second_tip_row = self.db.get_tip(tip_id="2")
-
-        # Creaing a Tip object
-        second_tip = Tip(self.db, str(second_tip_row["tip_id"]), second_tip_row)
-
-        print("getting second tip by id:")
-        print("tip_id:" + second_tip.db_id)
-        print("task_name: " + second_tip.task_name)
-        print("tip_text: " + second_tip.tip_text)
-        print("-----------------")
-
-        # Removing the table
-        self.db.drop_table(table_name="tips")
-        print("Dropped tips table")
+            # Removing the table
+            self.db.drop_table(table_name="tips")
+            print("Dropped tips table")
